+++ crates/harness-server/src/workflow_runtime_worker.rs
@@ -13,6 +13,7 @@ mod prompt_input_telemetry;
 mod prompt_packet;
 mod runtime_profile;
 mod server_merge;
+pub(crate) mod turn_engine;
 mod workspace;
 
 use crate::http::AppState;
