+++ crates/harness-server/src/workflow_runtime_worker/turn_engine/helpers.rs
@@ -1,0 +1,187 @@
+use harness_core::agent::StreamItem;
+use harness_core::types::{ThreadId, TurnId, TurnStatus};
+use harness_protocol::{notifications::Notification, notifications::RpcNotification};
+
+pub(crate) fn emit_runtime_notification(
+    notify_tx: &Option<crate::notify::NotifySender>,
+    notification_tx: &tokio::sync::broadcast::Sender<RpcNotification>,
+    notification: Notification,
+) {
+    crate::notify::emit(notify_tx, notification.clone());
+    let _ = notification_tx.send(RpcNotification::new(notification));
+}
+
+pub(crate) async fn persist_runtime_thread(
+    thread_db: &Option<crate::thread_db::ThreadDb>,
+    server: &crate::server::HarnessServer,
+    thread_id: &ThreadId,
+) {
+    if let Some(db) = thread_db {
+        if let Some(thread) = server.thread_manager.get_thread(thread_id) {
+            if let Err(err) = db.update(&thread).await {
+                tracing::warn!("thread_db persist failed during turn execution: {err}");
+            }
+        }
+    }
+}
+
+pub(crate) async fn process_stream_item(
+    server: &crate::server::HarnessServer,
+    thread_db: &Option<crate::thread_db::ThreadDb>,
+    notify_tx: &Option<crate::notify::NotifySender>,
+    notification_tx: &tokio::sync::broadcast::Sender<RpcNotification>,
+    thread_id: &ThreadId,
+    turn_id: &TurnId,
+    stream_item: StreamItem,
+) {
+    match stream_item {
+        StreamItem::ItemStarted { item } => {
+            if let Err(err) = server
+                .thread_manager
+                .add_item(thread_id, turn_id, item.clone())
+            {
+                tracing::warn!("failed to append stream item_started to turn: {err}");
+            } else {
+                persist_runtime_thread(thread_db, server, thread_id).await;
+            }
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::ItemStarted {
+                    turn_id: turn_id.clone(),
+                    item,
+                },
+            );
+        }
+        StreamItem::ItemCompleted { item } => {
+            if let Err(err) = server
+                .thread_manager
+                .add_item(thread_id, turn_id, item.clone())
+            {
+                tracing::warn!("failed to append stream item_completed to turn: {err}");
+            } else {
+                persist_runtime_thread(thread_db, server, thread_id).await;
+            }
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::ItemCompleted {
+                    turn_id: turn_id.clone(),
+                    item,
+                },
+            );
+        }
+        StreamItem::TokenUsage { usage } => {
+            if let Err(err) =
+                server
+                    .thread_manager
+                    .set_turn_token_usage(thread_id, turn_id, usage.clone())
+            {
+                tracing::warn!("failed to update turn token usage: {err}");
+            } else {
+                persist_runtime_thread(thread_db, server, thread_id).await;
+            }
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::TokenUsageUpdated {
+                    thread_id: thread_id.clone(),
+                    usage,
+                },
+            );
+        }
+        StreamItem::Error { message } => {
+            if let Err(err) = server.thread_manager.add_item(
+                thread_id,
+                turn_id,
+                harness_core::types::Item::Error { code: -1, message },
+            ) {
+                tracing::warn!("failed to append stream error item to turn: {err}");
+            } else {
+                persist_runtime_thread(thread_db, server, thread_id).await;
+            }
+        }
+        StreamItem::MessageDelta { text } => {
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::MessageDelta {
+                    turn_id: turn_id.clone(),
+                    text,
+                },
+            );
+        }
+        StreamItem::ToolOutputDelta { item_id, text } => {
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::ToolOutputDelta {
+                    turn_id: turn_id.clone(),
+                    item_id,
+                    text,
+                },
+            );
+        }
+        StreamItem::ApprovalRequest { id, command } => {
+            if let Err(err) = server.thread_manager.add_item(
+                thread_id,
+                turn_id,
+                harness_core::types::Item::ApprovalRequest {
+                    id: Some(id.clone()),
+                    action: command.clone(),
+                    approved: None,
+                },
+            ) {
+                tracing::warn!("failed to append approval request item to turn: {err}");
+            } else {
+                persist_runtime_thread(thread_db, server, thread_id).await;
+            }
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::ApprovalRequest {
+                    turn_id: turn_id.clone(),
+                    request_id: id,
+                    command,
+                },
+            );
+        }
+        StreamItem::Warning { message } => {
+            emit_runtime_notification(
+                notify_tx,
+                notification_tx,
+                Notification::Warning {
+                    turn_id: turn_id.clone(),
+                    message,
+                },
+            );
+        }
+        _ => {}
+    }
+}
+
+pub(crate) async fn mark_turn_failed(
+    server: &crate::server::HarnessServer,
+    thread_db: &Option<crate::thread_db::ThreadDb>,
+    notify_tx: &Option<crate::notify::NotifySender>,
+    notification_tx: &tokio::sync::broadcast::Sender<RpcNotification>,
+    thread_id: &ThreadId,
+    turn_id: &TurnId,
+    error: String,
+) {
+    if let Err(err) = server.thread_manager.fail_turn(thread_id, turn_id) {
+        tracing::warn!("failed to mark turn as failed: {err}");
+    } else {
+        persist_runtime_thread(thread_db, server, thread_id).await;
+    }
+    emit_runtime_notification(
+        notify_tx,
+        notification_tx,
+        Notification::TurnCompleted {
+            turn_id: turn_id.clone(),
+            status: TurnStatus::Failed,
+            token_usage: harness_core::types::TokenUsage::default(),
+        },
+    );
+    tracing::error!("turn failed: {error}");
+}
